Update Lumiere.java
First thing to do is to change it to scan for six things. X, Y, and Z. I need this to provide a place to put the color point. After that, I to scan for R, G, and B. Once I get the color, I can create a picture using Java's functions.